requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
numpy
pillow
opencv-python-headless
torch==2.3.1+cpu         --index-url https://download.pytorch.org/whl/cpu
torchvision==0.18.1+cpu  --index-url https://download.pytorch.org/whl/cpu
